Refactor admin panel
- separate reports/statistics to own view
- separate past events to own view
- move tags under admin
- sepate out events/applicatons logic for admin from regular controllers
to admin controllers
