+++ app/assets/stylesheets/application.css.scss
@@ -52,6 +52,7 @@ $topbar-height: 55px;
 @import "login.scss";
 @import "map.scss";
 @import "searchbar.scss";
+@import "admin.scss";
 
 *, *:before, *:after {
   box-sizing: border-box;
