+++ app/assets/stylesheets/events.scss
@@ -38,104 +38,3 @@
     font-weight: 600;
   }
 }
-
-.approve-button {
-  input[type="checkbox"] {
-    margin-left: 10px;
-  }
-}
-
-.admin-list,
-.tags-list {
-  list-style: none;
-  padding: 0;
-
-  li {
-    margin-bottom: 1em;
-    border-bottom: 3px solid #f1f1f1;
-    @media (min-width: $medium) {
-      display: flex;
-      flex-flow: row wrap;
-      justify-content: space-between;
-      border-bottom: none;
-      margin-bottom: 0;
-    }
-  }
-  .application-name {
-    flex: 0 0 40%;
-    white-space: nowrap;
-    overflow: hidden;
-    margin-right: 1em;
-  }
-  .application-mail {
-    flex-basis: 40%;
-    overflow: hidden;
-    margin-right: 1em;
-    flex-grow: 1;
-  }
-  p {
-    margin: 0 0 1em;
-  }
-  .admin-buttons {
-    display: inline-block;
-  }
-  .event-name {
-    flex-basis: 15em;
-    white-space: nowrap;
-    overflow: hidden;
-  }
-  .application-process {
-    flex-basis: 12em;
-  }
-  .application-deadline {
-    flex-basis: 11em;
-  }
-}
-
-.organizer-list,
-.applicant-list {
-  list-style: none;
-  padding: 0;
-
-  li {
-    margin-bottom: 1em;
-    border-bottom: 3px solid #f1f1f1;
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: space-between;
-    border-bottom: none;
-    margin-bottom: 0;
-    p:not(:last-child) {
-      margin-right: 1em;
-    }
-  }
-  p {
-    margin: 0 0 1em;
-  }
-  .event-name {
-    flex-grow: 1;
-    white-space: nowrap;
-    overflow: hidden;
-  }
-}
-
-ul.tags {
-  padding-left: 0;
-
-  li.tag {
-    list-style: none;
-    padding: 0.2em 0.5em;
-    border-radius: 3px;
-    display: inline-block;
-    margin: 0 0 0.5em;
-    &.programming-languages {
-      background-color: lighten($blue-bg, 30%);
-    }
-    &.languages {
-      background-color: lighten($green-bg, 20%);
-    }
-    &.fields-of-expertise {
-      background-color: lighten($yellow, 20%);
-    }
-  }
-}
