+++ app/helpers/events_helper.rb
@@ -1,5 +1,4 @@
 module EventsHelper
-
   def event_image(event)
     if event.logo && event.logo != ''
       image_tag event.logo, skip_pipeline: true
@@ -9,15 +8,11 @@ module EventsHelper
   end
 
   #Used in events/edit and events/show:
-
   def breadcrumb_locals_according_to_user_status
     if current_user && @event.organizer_id == current_user.id
       { your_events: t('.your_events') }
-    elsif current_user && current_user.admin?
-      { admin: t('.admin_link') }
     else
       { events: t('.events_link') }
     end
   end
-
 end
