+++ app/views/admin/_admin_navigation.html.erb
@@ -1,0 +1,9 @@
+<nav class="secondary-navigation" role="navigation">
+  <p>Admin</p>
+  <ul>
+    <li><%= navbar_link_to "Events", admin_events_path %></li>
+    <li><%= navbar_link_to "Past Events", past_admin_events_path %></li>
+    <li><%= navbar_link_to "Tags", admin_tags_path %></li>
+    <li><%= navbar_link_to "Reports", reports_admin_events_path %></li>
+  </ul>
+</nav>
