+++ app/views/admin/events/edit.html.erb
@@ -1,0 +1,19 @@
+<% provide :title, "Admin - Edit #{@event.name}" %>
+
+<%= render partial: "shared/breadcrumbs", locals: {admin: "Admin", edit_event: "Edit Event"} %>
+
+<h1 class="page-title"><span class="page-title-firstword">Edit</span> <%= @event.name %></h1>
+
+<% if @event.errors.any? %>
+  <div class="error">
+    <p><%= pluralize(@event.errors.count, "error", "errors") %></p>
+
+    <ul>
+      <% @event.errors.messages.each do |field, messages| %>
+        <li><%= @event.class.human_attribute_name(field) %> <%= join_messages(messages) %></li>
+      <% end %>
+    </ul>
+  </div>
+<% end %>
+
+<%= render partial: "events/form", locals: {submit_tag_text: "Submit Changes", form_url: admin_event_path(@event)} %>
