+++ app/views/admin/events/reports.html.erb
@@ -1,0 +1,46 @@
+<% provide :title, "Admin - Reports" %>
+
+<%= render partial: "admin/admin_navigation" %>
+
+<%= render partial: "shared/breadcrumbs", locals:{ title: "Reports", admin: "Admin" } %>
+
+<h1>Reports</h1>
+
+<section class="box">
+  <h2><%= "Downloads" %></h2>
+  <%= link_to "Download events report", admin_events_path(format: :csv), class: "btn btn-save"%>
+  <%= link_to "Download annual events report", annual_admin_events_path(format: :csv), class: "btn btn-save"%>
+</section>
+
+<div class="box">
+  <h2><%= "Statistics" %></h2>
+  <ul class="admin-list">
+    <li>Total conferences:<strong><%= @events.count %></strong></li>
+    <li><%= "Conferences since #{Time.zone.now.year}:" %><strong><%= @events.created_current_year.count %></strong></li>
+    <li>Total organizers:<strong><%= @total_organizers %></strong></li>
+    <li class="border">New users in the last 30 days:<strong><%= @new_users.count %></strong></li>
+  </ul>
+  <ul class="admin-list">
+    <li>
+      Most conferences
+      <strong>
+        <ol>
+          <% @country_rank.each.with_index do |country, index| %>
+          <% break if index >= 3 || index == @country_rank.length %>
+          <li><%= index + 1 %>. <%= @country_rank[index][0] %> (<%= @country_rank[index][1] %>)</li>
+        <% end %>
+        </ol>
+      </strong>
+    </li>
+  </ul>
+  <div class="map-statistics" id="map-statistics">
+    <div class="map-statistics-content">
+      <div class="map-legend">
+        <span>&nbsp;</span>
+        Countries that hosted events
+      </div>
+      <div id="map_events" style="position: relative; width: 100%; height: 600px;"></div>
+      <%= javascript_tag "new CustomizedMap(#{@countries_statistics});" %>
+    </div>
+  </div>
+</div>
