--- app/views/admin_applications/show.html.erb
@@ -1,61 +1,0 @@
-<% provide :title, t('.title') %>
-
-<%= render partial: "shared/breadcrumbs", locals:{ admin: t('.admin_link'), event: @event, application_by: t('.application_by', applicant_name: @application.name) } %>
-
-<h1 class="box-title"><%=t('.application_for', event_name: @event.name)%></h1>
-
-<section class="box">
-  <p class="detail-pair">
-    <strong>
-      <%= t('.attendee_info_1', event_name: @event.name) %>
-    </strong>
-    <%= @application.attendee_info_1 %>
-  </p>
-
-  <p class="detail-pair">
-    <strong><%= t('.attendee_info_2') %></strong>
-    <%= @application.attendee_info_2 %>
-  </p>
-
-  <p class="detail-pair">
-    <strong><%= t(".name") %></strong>
-    <%= @application.name %>
-  </p>
-
-  <p class="detail-pair">
-    <strong><%= t(".email") %></strong>
-    <%= @application.email %>
-  </p>
-
-  <% if @application.visa_needed %>
-    <p class="detail-pair">
-      <strong><%= t(".visa_need") %></strong>
-    </p>
-  <% end %>
-</section>
-
-<div class="box">
-  <% if @event.application_process == "selection_by_travis" %>
-    <p><%=t('.application_status', application_status: @application.status)%></p>
-    <% if @application.event.closed? && !@event.past? %>
-      <% if @application.status == 'pending' %>
-        <%= link_to t(".approve"), approve_event_application_path(@application.event_id, @application.id),
-            method: :post, class: "btn btn-save approve-application",
-            title: t(".approve").downcase %>
-        <%= link_to t(".reject"), reject_event_application_path(@application.event_id, @application.id),
-         method: :post, class: "btn btn-external reject-application",
-         title: t(".reject").downcase %>
-      <% elsif @application.status == 'approved' || @application.status == 'rejected' %>
-        <%= link_to t(".revert"), revert_event_application_path(@application.event_id, @application.id),
-               method: :post, class: "btn btn-edit revert-application",
-               title: t(".revert").downcase %>
-      <% end %>
-    <% end %>
-  <% end %>
-  <%= link_to t('.delete_application'), event_application_path(@event.id, @application.id),
-      class: "btn btn-delete",  method: :delete,
-      data: { confirm: t(".are_you_sure") } %>
-</div>
-<p>
-  <%= link_to t(".show_event_details"), event_path(@event.id) %>
-</p>
