--- app/views/admin_events/index.html.erb
@@ -1,134 +1,0 @@
-<% provide :title, t('.title') %>
-
-<%= render partial: "shared/breadcrumbs", locals:{ title: t('.title') } %>
-
-<h1><%=t('.title')%></h1>
-
-<%= render partial: 'statistics' %>
-
-<p>
-  <%= link_to t('.download_events_report'), admin_path(format: :csv),
-    class: "btn btn-save"%>
-
-  <%= link_to t('.download_annual_events_report'), admin_annual_path(format: :csv),
-    class: "btn btn-save"%>
-</p>
-
-
-<div class="box">
-  <h2>Manage "Selection by Diversity Tickets"</h2>
-  <%= form_for @application_process_options_handler, url: update_process_path do |form| %>
-    <%= form.check_box :selection_by_dt_enabled, value: selection_by_dt_enabled?  %>
-      Enable selection by Diversity Tickets?
-    <div class="submit-field">
-      <%= form.submit(class: "btn btn-save", value: 'Save', data: {
-        confirm: selection_by_dt_enabled? ? 'Are you sure you want to disable selection by Diversity Tickets?' : 'Are you sure you want to enable selection by Diversity Tickets?'
-      }) %>
-    </div>
-  <% end %>
-</div>
-
-<% @categorized_events.each do |heading, events| %>
-  <div class="box">
-    <h2><%= heading %></h2>
-
-    <div class="admin">
-      <ul class="admin-list">
-        <% if events.empty? %>
-        <p><%= t(".no_events_to_approve")%></p>
-      </ul>
-        <% elsif heading == "Approved events" %>
-          <% @approved_events_deadline.each do |heading, events| %>
-            <% unless events.empty? %>
-              <div class="box">
-                <h3><%= heading %></h3>
-                <% events.each do |event| %>
-                <li>
-                  <p class="event-name">
-                    <span class="conflogo--small"><%= event_image event %></span>
-                    <%= link_to event.name, event_admin_path(id: event.id), title: event.name %>
-                  </p>
-                  <p class="application-process">
-                    <%= event.application_process.humanize %>
-                  </p>
-                  <p class="application-deadline">
-                    <%= format_date(event.deadline) %>
-                  </p>
-                  <div class="admin-buttons">
-                    <%= link_to approve_admin_event_path(event),
-                        method: :post, class: "icon tooltip-large",
-                        title: t('.unapprove') do %>
-                        <span class="icon-unapprove" aria-label="unapprove icon"></span>
-                      <% end %>
-
-                    <%= link_to edit_event_path(id: event.id), class: "icon tooltip", title: t('.edit') do %>
-                      <span class="icon-edit" aria-label="edit icon"></span>
-                    <% end %>
-                    <%= link_to url_for(action: :destroy, id: event.id), method: :delete,
-                        data: { confirm: t('.are_you_sure')}, class: "icon tooltip", title: t('.delete') do %>
-                      <span class="icon-delete" aria-label="trashcan icon"></span>
-                    <% end %>
-                  </div>
-                  <p>
-                    <%= download_csv_link(event) %>
-                  </p>
-                </li>
-                <% end %>
-              </div>
-            <% end %>
-          <% end %>
-        <% else %>
-        <% events.each do |event| %>
-          <li>
-            <p class="event-name">
-              <span class="conflogo--small"><%= event_image event %></span>
-              <%= link_to event.name, event_admin_path(id: event.id), title: event.name %>
-            </p>
-            <p class="application-process">
-              <%= event.application_process.humanize %>
-            </p>
-            <p class="application-deadline">
-              <%= format_date(event.deadline) %>
-            </p>
-            <div class="admin-buttons">
-              <% if event.unapproved %>
-                <%= form_for :approve,
-                      url: approve_admin_event_path(event),
-                      html: { method: :post } do |form| %>
-                    <span>
-                      <span class="icon tooltip" title="tweet">
-                      <%= form.check_box :tweet, checked: true %></span>
-                      <span class="icon-twitter icon tooltip" aria-label="twitter icon"></span>
-                      <span class="icon tooltip" title="approve" >
-                      <%= form.submit(class: "icon-approve", title: "approve") %></span>
-                    <% end %>
-              <% else %>
-                <%= link_to approve_admin_event_path(event),
-                    method: :post, class: "icon tooltip-large",
-                    title: t('.unapprove') do %>
-                    <span class="icon-unapprove" aria-label="unapprove icon"></span>
-                  <% end %>
-              <% end %>
-
-              <%= link_to edit_event_path(id: event.id), class: "icon tooltip", title: t('.edit') do %>
-
-                <span class="icon-edit" aria-label="edit icon"></span>
-              <% end %>
-              <%= link_to url_for(action: :destroy, id: event.id), method: :delete,
-                  data: { confirm: t('.are_you_sure')}, class: "icon tooltip", title: t('.delete') do %>
-                <span class="icon-delete" aria-label="trashcan icon"></span>
-              <% end %>
-            </div>
-            <p>
-              <% if event.approved %>
-              <%= link_to t('.download_csv'), event_admin_path(id: event.id, format: :csv),
-                  class: "btn btn-save", title: t('.download_data') %>
-              <% end %>
-            </p>
-          </li>
-          <% end %>
-        <% end %>
-      </ul>
-    </div>
-  </div>
-<% end %>
