+++ app/views/admin_mailer/passed_event_deadline.text.erb
@@ -6,4 +6,4 @@ Application Process: <%= @event.application_process.humanize %>
 Follow the link below to select the
 winners:
 
-<%= event_admin_url(@event.id, locale: "en") %>
+<%= admin_event_url(@event.id, locale: "en") %>
