+++ app/views/admin_mailer/submitted_event.text.erb
@@ -8,4 +8,4 @@ Application Process: <%= @event.application_process.humanize %>
 
 Follow the link below to view details:
 
-<%= event_admin_url(@event.id, locale: "en") %>
+<%= admin_event_url(@event.id, locale: "en") %>
