+++ app/views/admin_mailer/upcoming_event_deadline.text.erb
@@ -5,4 +5,4 @@ Application Process: <%= @event.application_process.humanize %>
 
 Follow the link below to view details:
 
-<%= event_admin_url(@event.id, locale: "en") %>
+<%= admin_event_url(@event.id, locale: "en") %>
