+++ app/views/events/show.html.erb
@@ -5,13 +5,12 @@
 
 <%= render partial: "shared/breadcrumbs", locals: Hash[key => value, event: @event, approval_pending: t('.approval_pending')] %>
 
-
 <h1><span class="conflogo--medium"><%= event_image @event %></span><%= @event.name %></h1>
 
 <div class="two-columns">
   <%= render partial: "events/event" %>
 
-  <% if current_user && (@event.organizer_id == current_user.id || current_user.admin?) %>
+  <% if current_user && @event.organizer_id == current_user.id %>
     <% if @event.closed? %>
       <section class="box">
         <h2 class="box-title">
