+++ app/views/shared/_breadcrumbs.html.erb
@@ -1,14 +1,11 @@
 <p class="breadcrumb">
-  <%= link_to t('.home_link'), root_path %> >
-
-  <% if local_assigns.has_key? :title %>
-    <%= title %>
-
-  <% elsif local_assigns.has_key? :admin %>
+  <% if local_assigns.has_key? :admin %>
     <%= link_to admin, admin_path %> >
-    <% if local_assigns.has_key? :application_by %>
-      <%= link_to event.name, event_admin_path(event.id) %> >
-      <%= application_by %>
+    <% if local_assigns.has_key? :title %>
+      <%= title %>
+    <% elsif local_assigns.has_key? :application_by %>
+      <%= link_to event.name, admin_event_path(event.id) %> >
+      <%= "#{application_by} #{applicant_name}" %>
     <% elsif local_assigns.has_key? :edit_event %>
       <%= edit_event %>
     <% else %>
@@ -17,42 +14,48 @@
         <%= approval_pending %>
       <% end %>
     <% end %>
+  <% else %>
+    <%= link_to t('.home_link'), root_path %> >
 
-  <% elsif local_assigns.has_key? :events %>
-    <%= link_to events, events_path %> >
-    <% if local_assigns.has_key? :edit_event %>
-      <%= edit_event %>
-    <% elsif local_assigns.has_key? :approval_pending %>
-      <%= event.name %>
-      <% if !event.approved? %>
-        <%= approval_pending %>
-      <% end %>
-    <% else %>
-      <%= link_to event.name, event_path(event) %> >
-      <% if local_assigns.has_key? :application_for %>
-        <%= application_for %>
+    <% if local_assigns.has_key? :events %>
+      <%= link_to events, events_path %> >
+      <% if local_assigns.has_key? :edit_event %>
+        <%= edit_event %>
+      <% elsif local_assigns.has_key? :approval_pending %>
+        <%= event.name %>
+        <% if !event.approved? %>
+          <%= approval_pending %>
+        <% end %>
       <% else %>
-        <%= application %>
+        <%= link_to event.name, event_path(event) %> >
+        <% if local_assigns.has_key? :application_for %>
+          <%= application_for %>
+        <% else %>
+          <%= application %>
+        <% end %>
       <% end %>
-    <% end %>
 
-  <% elsif local_assigns.has_key? :your_applications %>
-    <%= link_to your_applications, user_applications_path(current_user.id) %> >
-    <% if local_assigns.has_key? :draft_for %>
-      <%= draft_for %>
-    <% else %>
-      <%= application_for %>
-    <% end %>
-
-  <% elsif local_assigns.has_key? :your_events %>
-    <%= link_to your_events, user_path(current_user) %> >
-    <% if local_assigns.has_key? :event %>
-      <%= event.name %>
-      <% if !event.approved? %>
-        <%= approval_pending %>
+    <% elsif local_assigns.has_key? :your_applications %>
+      <%= link_to your_applications, user_applications_path(current_user.id) %> >
+      <% if local_assigns.has_key? :draft_for %>
+        <%= draft_for %>
+      <% else %>
+        <%= application_for %>
       <% end %>
-    <% else %>
-      <%= edit_event %>
+
+    <% elsif local_assigns.has_key? :your_events %>
+      <%= link_to your_events, user_path(current_user) %> >
+      <% if local_assigns.has_key? :event %>
+        <%= event.name %>
+        <% if !event.approved? %>
+          <%= approval_pending %>
+        <% end %>
+      <% else %>
+        <%= edit_event %>
+      <% end %>
+
+    <% elsif local_assigns.has_key? :title %>
+      <%= title %>
     <% end %>
   <% end %>
 </p>
