+++ config/locales/en.yml
@@ -63,73 +63,6 @@ en:
               blank: "can't be blank"
               invalid: "should be a full URL, e.g. https://myconference.com"
 
-  admin_applications:
-    application_approved: "%{applicant_name}'s application has been approved!"
-    application_pending: "%{applicant_name}'s application has been changed to pending"
-    application_rejected: "%{applicant_name}'s application has been rejected"
-    show:
-      admin_link: "Admin"
-      application_by: "Application by %{applicant_name}"
-      application_for: "Application for %{event_name}"
-      application_status: "Application status: %{application_status}"
-      approve: "Approve"
-      are_you_sure: "Are you sure you want to delete this application?"
-      attendee_info_1: "Why do you want to attend %{event_name} and what especially do you look forward to learning?"
-      attendee_info_2: "Please share with us why you're applying for a diversity ticket."
-      delete_application: "Delete application"
-      email: "Email"
-      home_link: "Home"
-      name: "Name"
-      reject: "Reject"
-      revert: "Revert"
-      show_event_details: 'Show event details'
-      title: "Admin"
-      visa_need: "I need a Visa."
-
-  admin_events:
-    event_approved: "%{event_name} has been approved!"
-    event_unapproved: "%{event_name} has been unapproved"
-    index:
-      are_you_sure: "Are you sure?"
-      approve: "approve"
-      edit: "edit"
-      delete: "delete"
-      download_annual_events_report: "Download annual events report"
-      download_csv: "Download CSV"
-      download_data: "Download data"
-      download_events_report: "Download events report"
-      no_events_to_approve: "We currently don't have any events to approve"
-      title: "Admin page"
-      unapprove: "unapprove"
-    show:
-      admin_link: "Admin"
-      applications: "Applications"
-      approve: "Approve"
-      approve_event: "Approve event"
-      are_you_sure: "Are you sure?"
-      delete: "Delete"
-      delete_event: "Delete event"
-      download_csv: "Download CSV"
-      download_data: "Download data"
-      edit_event: "Edit event"
-      home_link: "Home"
-      organizer_details: "Organizer details"
-      organizer_email: "Email"
-      organizer_name: "Name"
-      title: "Admin"
-      tweet_event: "Tweet it"
-      reject: "Reject"
-      revert: "Revert"
-    statistics:
-      conferences_since: "Conferences since "
-      countries_hosted_events: "Countries that hosted events"
-      most_conferences: "Most conferences"
-      new_users_last_30_days: "New users in the last 30 days:"
-      total_conferences: "Total conferences:"
-      total_organizers: "Total organizers:"
-      show_statistics_map: "Show statistics map"
-      statistics: "Statistics"
-
   applications:
     already_applied: "You have already applied for %{event_name}"
     application_deleted: "You cannot view your application as the event you applied for has been removed from Diversity Tickets"
@@ -188,7 +121,6 @@ en:
       home_link: "Home"
       title: "New application"
     show:
-      admin_link: "Admin"
       application_by: "Application by %{applicant_name}"
       application_for: "Application for %{event_name}"
       approve: "Approve"
@@ -337,7 +269,6 @@ en:
       your_application: "Your application"
       your_draft: "Your draft"
     edit:
-      admin_link: "Admin"
       edit: "Edit"
       edit_event: "Edit event"
       events_link: "Events"
@@ -381,7 +312,6 @@ en:
       submit_your_event: "Submit your event"
       title: "Preview event"
     show:
-      admin_link: "Admin"
       applications_closed: "Sorry, applications for %{event_name} are closed."
       apply_at: "Apply at %{event_name}"
       apply_for_ticket: "Apply for a ticket"
@@ -574,12 +504,6 @@ en:
       people: "People"
       tech: "Tech"
       security: "Security"
-    index:
-      are_you_sure: "Are you sure?"
-      title: "Tags management"
-    tag_created: "Tag %{tag_name} was successfully created."
-    tag_problem: "There was a problem creating the new tag."
-
 
   users:
     account_deleted: "Your account has been deleted successfully."
