+++ config/locales/fr.yml
@@ -63,73 +63,6 @@ fr:
               blank: "obligatoire"
               invalid: "doit être une URL complète, comme https://maconference.fr"
 
-  admin_applications:
-    application_approved: "La candidature de %{applicant_name} a été approuvée"
-    application_pending: "La candidature de %{applicant_name} est en attente"
-    application_rejected: "La candidature de %{applicant_name} a été rejetée"
-    show:
-      admin_link: "Admin"
-      application_by: "Candidature de %{applicant_name}"
-      application_for: "Candidature pour %{event_name}"
-      application_status: "Statut de la candidature : %{application_status}"
-      approve: "Approuver"
-      are_you_sure: "Êtes-vous sûr·e de vouloir supprimer cette candidature ?"
-      attendee_info_1: "Pourquoi souhaitez-vous participer à %{event_name} et que souhaitez-vous tout particulièrement y apprendre ?"
-      attendee_info_2: "Qu'est-ce qui vous rend éligible à un billet diversité ? (1-2 phrases suffisent, ne partagez que ce avec quoi vous êtes à l'aise)"
-      delete_application: "Supprimer la candidature"
-      email: "Courriel"
-      home_link: "Accueil"
-      name: "Nom"
-      reject: "Rejeter"
-      revert: "Annuler"
-      show_event_details: "Afficher les détails de l'événement"
-      title: "Admin"
-      visa_need: "J'ai besoin d'un visa."
-
-  admin_events:
-    event_approved: "%{event_name} a été approuvé !"
-    event_unapproved: "%{event_name} n'a pas été approuvé."
-    index:
-      are_you_sure: "Êtes-vous sûr·e ?"
-      approve: "Approuver"
-      edit: "Modifier"
-      delete: "Supprimer"
-      download_annual_events_report: "Télécharger le rapport annuel des événements"
-      download_csv: "Télécharger CSV"
-      download_data: "Télécharger les données"
-      download_events_report: "Télécharger le rapport des événements"
-      no_events_to_approve: "Nous n'avons actuellement aucun événement à approuver."
-      title: "Page d'admin"
-      unapprove: "refuser"
-    show:
-      admin_link: "Admin"
-      applications: "Candidature"
-      approve: "Approuver"
-      approve_event: "Approuver l'événement"
-      are_you_sure: "Êtes-vous sûr·e ?"
-      delete: "Supprimer"
-      delete_event: "Supprimer l'événement"
-      download_csv: "Télécharger CSV"
-      download_data: "Télécharger les données"
-      edit_event: "Modifier l'événement"
-      home_link: "Accueil"
-      organizer_details: "Détails de l'organisateur"
-      organizer_email: "Courriel"
-      organizer_name: "Nom"
-      title: "Admin"
-      tweet_event: "Tweetez le"
-      reject: "Rejeter"
-      revert: "Annuler"
-    statistics:
-      conferences_since: "Événements depuis "
-      countries_hosted_events: "Pays hôtes"
-      most_conferences: "Le + d'événements"
-      new_users_last_30_days: "Nouveaux inscrits les 30 derniers jours:"
-      total_conferences: "Nombre total d'événements:"
-      total_organizers: "Nombre total d'organisateurs:"
-      show_statistics_map: "Afficher la carte statistique"
-      statistics: "Statistiques"
-
   applications:
     already_applied: "Vous avez déjà postulé à une place pour %{event_name}"
     application_deleted: "Vous ne pouvez pas voir votre candidature car l'événement en question a été retiré de Diversity Tickets."
@@ -188,7 +121,6 @@ fr:
       home_link: "Accueil"
       title: "Nouvelle candidature"
     show:
-      admin_link: "Admin"
       application_by: "Candidature de %{applicant_name}"
       application_for: "Candidature pour %{event_name}"
       approve: "Approuver"
@@ -337,7 +269,6 @@ fr:
       your_application: "Votre candidature"
       your_draft: "Votre brouillon"
     edit:
-      admin_link: "Admin"
       edit: "Modifier"
       edit_event: "Modifier l'événement"
       events_link: "Événements"
@@ -381,7 +312,6 @@ fr:
       submit_your_event: "Soumettre votre événement"
       title: "Prévisualiser l'événement"
     show:
-      admin_link: "Admin"
       applications_closed: "Désolé, les candidatures pour %{event_name} sont fermées."
       apply_at: "Postuler à %{event_name}"
       apply_for_ticket: "Faire une demande de billet"
@@ -526,11 +456,6 @@ fr:
       people: "People"
       tech: "Tech"
       security: "Sécurité"
-    index:
-      are_you_sure: "Êtes-vous sûr·e ?"
-      title: "Gestion des étiquettes"
-    tag_created: "L'étiquette %{tag_name} a été créée avec succès."
-    tag_problem: "Un problème est survenu durant la création de la nouvelle étiquette."
 
   users:
     account_deleted: "Votre compte a été supprimé avec succès."
