+++ test/integration/admin_event_details_test.rb
@@ -2,7 +2,6 @@ require 'test_helper'
 include ApplicationHelper
 
 feature 'Admin Event Details' do
-
   def setup
     make_application_process_options_handler
     @admin = make_admin
@@ -35,7 +34,6 @@ feature 'Admin Event Details' do
     visit admin_event_path(@event.id)
 
     assert_not page.has_content?('Mary')
-
   end
 
   test 'shows submitted applications on event details page for admin' do
@@ -90,7 +88,7 @@ feature 'Admin Event Details' do
 
     page.first('a.btn.btn-save.approve-application').click
 
-    assert page.has_content?("Peter's application has been approved!")
+    assert page.has_content?("Peter's application has been approved")
 
     assert page.has_content?("Approved applications (2)")
     assert page.has_content?("Pending applications (1)")
@@ -134,10 +132,10 @@ feature 'Admin Event Details' do
 
     visit admin_event_path(@event.id)
 
-    assert page.has_selector?("input[type=submit][value='Approve event']")
-    click_button("Approve event")
+    assert page.has_selector?("input[type=submit][value='Approve']")
+    click_button("Approve")
 
-    assert page.has_content?("#{@event.name} has been approved!")
+    assert page.has_content?("#{@event.name} has been approved")
   end
 
   test 'shows correct number of approved tickets' do
